Fix the inconsistent return value of notifications
The API returned a different data structure when there are no notifications to show compared to when there are items. This causes the JS-API to break, because it always expects the later data structure.

This is a regression of ec1a4a80d3a8f89105d6012ec857a7115f41c97e

See https://www.woltlab.com/community/thread/296174-benachrichtigungen-werden-nicht-mehr-angezeigt/