Explicitly handle a `null` value in UserOptionOutput
userOptions are commonly nullable, because no default value can be defined for
the `TEXT` column type family.

`null` needs to be handled explicitly, because PHP 8.1 deprecated passing
`null` to non-nullable string parameters of native functions. As an example
without this change, `null` could be passed to `StringUtil::trim()` in
`MessageUserOptioNOutput` which in turn passes it to `preg_replace()`.