Return `TIME_NOW` in Session::getLastActivityTime() for the current session
Since b1adc239cb469a2c648d56d5e98e57c8a41d9b2c the lastActivityTime is only
updated once per minute to reduce the write traffic on the session table. This
might lead to the current session not appearing at the top in account security.

Fix this by returning the current time in `getLastActivityTime()` if the
session in question is the active session, because the current request is the
latest activity by definition.