Show WCF_N in ACP's index.tpl only when not 1
Since e3fd38fd7ef7e3bb1dbb871fdd6786c7009aa36d it is no longer possible to
configure WCF_N to anything other than `1`. Only updated installations might
still have a value that is not `1` and even then it should be rare.

Do not show the number, unless it is non-standard to not confuse the reader
with information that are not relevant.