Preserve the sequence numbers when cloning a package node
This change is safe, the only place where the sequence number is used is within
the `ORDER BY` when reading the nodes for the current request.

Preserving the sequence numbers makes debugging a little easier, because steps
that were created using a split are immediately recognizable by their sequence
numbers not starting at 0.