Add proper parameter types to UserStorageHandler
Previously attempting to store an array in the user storage without manually
serializing it appeared to succeed, as the `->update()` call accepted it and
stored it in the `->log`. However the actually persisting would fail during
shutdown, as the array would be implicitly converted to a string, thus causing
a PHP Warning to be emitted.

Add parameter types to ensure the error is immediately detected when calling
`->update()`, as errors during shutdown are very hard to debug.

Return types are not added, as UserStorageHandler is not final and thus
technically child classes could exist. Adding types would break these child
classes.

see https://www.woltlab.com/community/thread/297525-array-to-string-conversion-fehlermeldung/