Add a relocatable global page overlay container
The dialog element as well as its backdrop are implemented as fullscreen elements in browser. This will cause them to be _always_ at the top layer.

This breaks any content inside the dialog that effectively lives outside of the dialog, such as tooltips or drop-down menus.

Dynamically moving the container in and out of the dialog will preserve the stacking context and allow these elements to work normally.