Call `initSecurityToken()` where previously `defineConstants()` was called
Since 011293191d36552a12c801247297410c188e9030 the XSRF-TOKEN was only set when
->getSecurityToken() was called, effectively only when validating and when
embedding the token in an URL (which should not be done in the first place).

Thus a valid cookie might not exist for a request. In the future the cookie /
XSRF Token handling should move into a middleware as much as possible.