Stop providing the superglobals and `$tpl.now` in templates
Templates have no business of accessing the super globals directly. This type
of business logic needs to happen in proper PHP code. This is especically true
with PSR-15 controllers which phase out the use of superglobals even for PHP.

The `$tpl.now` value is more easily accessed using the `TIME_NOW` constant.