Inline WoltLabSuite/Core/I18n/Plural in Template module
In c0b2cbfb163d298bbfffa6f3824a287fd4e273a5 the vast majority of the Plural
module contents has already been removed in favor of Intl.PluralRules, only the
template functionality remained. This commit inlines the functionality into the
Template module to allow removing the module entirely.

The differences in Template.grammar.js are larger than expected, apparently
some kind of autoformatter was once applies to the file by accident. I have
verified with `git diff --word-diff` that the renaming of the function call was
the only real change in the file.