Return PSR-7 response in AbstractOauth2Action::initiate()
Also technically a small break if some code calls `->initiate()` manually and
expects it to `never` return due to the `exit;`. Any classes overriding
`initiate()` will notice due to the new return type of the previous commit, any
other manual uses are very unlikely and not how AbstractOauth2Action is
supposed to be used.