Improve performance of CryptoUtil::getValueFromSignedString()
Previously the method required to b64 decode the value twice, once for the
validation in `validateSignedString()` and once more to actually decode it.
This was fairly slow, due to the constant time encoder being used for decoding,
causing the b64 decoder to account for more almost 1% of time spent for simple
requests, now it's down to ~0.6% of the reduced total.

The change is safe, because there is no way for valid return value to be
confused with an invalid return value, due to entirely distinct types being
used (string vs null). Thus `validateSignedString()` can simply delegate to
`getValueFromSignedString()`.