Use absolute paths in require() calls during script init
Relative paths in `require()` are relative to the accessed “entrypoint” (i.e.
not the script that calls `require()`). By using absolute paths the requiring
is more resilient and more specifically the `acp/global.php` no longer depends
on an application correctly configuring `RELATIVE_WCF_DIR`, removing one of the
remaining users of the `RELATIVE_*` constants.