Update the pending packages after the plugin nodes were built
Previously the `buildPluginNodes()` step in PackageInstallationNodeBuilder
would attempt to find update instructions for the *target* version, instead of
the source version. We must move the update of the pending packages below this
step to fix this. It also makes sense, because only once the plugin nodes have
been built (and thus executed) the package logically is on the new version.

see 5df3e65185d4d3f1be7464b1999b949946a5c765