Fix the access to the prototype during the object initialization
`_readField()` is implicitly called in the constructor of the parent class. This is effectively an attempt to access `this` while `super()` is still running which is a silent violation of the specs.

The `_fields` property in line 15 is actually implicitly declared in the constructor *after* `super()` has completed.

JavaScript’s technical debts strike again.