Overhaul weird code for the loading of responses
It is handled here the case when a user jumps via direct link to a comment that is not visible by default. In this case no responses are shown by default (which is intended).

The code worked correctly but was difficult to understand because the date for the last reply was set to the date of the first reply. Also, the SQL query was unnecessary because we can set the `lastResponseTime` passed via AJAX hard-coded to `1` (`0` would trigger a validation error).