Generate cleaner output in EmailGrammar::encodeQuotedPrintableHeader()
This updates the quoted-printable header encoder to not unnecessarily apply
quoted-printable encoding:

Previously if the full input would not fit into the maximum length of 70
characters, encoding would be enforced to keep things simple on the encoder
side, but resulting in overly bloated headers. This is especially visible with
the 'Subject' header of notification mails.

Now the quoted-printable encoding will only be applied if there actually are
any special characters. If the full input exceeds the maximum length without
containing special characters, it will be correctly broken at the natural space
boundaries.

Resolves #4970