Add return type to IBBCode::getParsedTag()
This improves debugging if a BBCode implementation is buggy and accidentally
does not return something, thus implicitly returning `null` without the type.
This would then (much) later throw an error, because `null` is passed to a
function that expects a string.

see https://www.woltlab.com/community/thread/298693-mb-strpos-passing-null-to-parameter-1-haystack-of-type-string-is-deprecated/