implement `TI18nDatabaseObjectAction`
`TI18nDatabaseObjectAction` provides two functions to be used in `AbstractDatabaseObjectAction::create()` and `AbstractDatabaseObjectAction::update()` when using the `FormBuilder`-API:
`prefillI18nColumn()` handles missing values for columns, that don't have a default value and can't be `NULL`. This function is meant to be used for this case only, not as a generic snippet.
`saveI18nColumn()` handles the final saving part. It detects whether i18n values are given for a property, saves them and updates the object-property's value to the language item.
This trait currently supports any i18n-properties that match `$propertyName_i18n`; both functions have to be called explicitly per property to prevent wildcard executions.