Include locale in `accept-language` for unfurl requests
It seems unlikely that a website would offer a region specific localized
version, but we have the necessary locale information readily available. So we
can just include it within the `accept-language` header for completeness.

Now the locale variant will be preferred. If the language code differs (i.e. a
specific locale is configured) it will be requested with a slightly lower
priority and lastly there still is the fallback to anything with a low
priority.

see f4d5b62d6777455f69d4d59b6976a6f86484c2ff