Remove `WCF_N` check in LanguageFactory::loadCache()
As explained in the previous commit's message, this is a very non-obvious
behavior, because checking for WCF_N differs from the check that is usually
used to test for WCFSetup: `!PACKAGE_ID`.

It's no longer necessary and thus removed. Anything that relies on
LanguageFactory too early in the request will then get an error, instead of
silently breaking LanguageFactory for the remainder of the request.