Fix the `addEventListener` definition for custom elements
The previous definition violated the liskov substitution principle, because not
all valid calls for the parent type `HTMLElement` are valid calls on the custom
element. Specifically it was not allowed to e.g. bind a `click` (or any other
“standard”) event onto a `woltlab-core-dialog` element.

This error was reported when consuming the `.d.ts` definitions in a TypeScript
project with `strict: true`.

Also the `addEventListener` was needlessly emitted as a real function that just
delegates to `super.addEventListener()`.

Fix both issues by moving the function signature definition into an `interface`
(to avoid the code generation) and by intersecting our more specific definition
with the `HTMLElement["addEventListener"]` definition to inherit all signatures
from `HTMLElement`.

see microsoft/TypeScript#52916