Stop needlessly invoking the autoloader to `<wcfNode-*>` tags
These node names are randomly generated and will never match an existing class.
In fact the tag name transformation did not even result in valid classnames in
the majority of cases, because the randomly generated ID might start with a
number, thus including a hyphen in the classname.

Stop needlessly invoking the autoloader, and thus accessing the disk, for these
tags.