Remove FirstTimeSetupForm.class.php on upgrade
Upon performing a software upgrade, an issue has been identified whereby the user becomes constrained within the previous iteration of the First Time Setup page. Despite the presence of a new version, the user is presented with the outdated version due to the prioritization of the FirstTimeSetupForm over the recently added FirstTimeSetupAction. This issue stems from the oversight of not removing the previous controller and template during the _upgrade_ from version 5.5, resulting in a conflict with the updated software. In order to resolve this issue, it is necessary to remove the outdated components using the *delete-PiPs. Failure to do so may result in users being unable to properly utilize the updated software.