Use intl Collator to sort groups in EditorGetMentionSuggestionsAction
The Collator is able to correctly handle accented characters, as it does not
sort based on the byte values of the UTF-8 representation.

One example would be Latin Capital Letter A with Grave (U+00C0) and Latin
Capital Letter A with Acute (U+00C1). The acute is correctly ordered before the
Grave, but the unicode number of the letter with the acute is the higher one.

see #4672
see 3db0f431b8db3aebb11ec9ac3d69a6199143b738