Fix the animation of dialogs in Safari
Safari executes the resize observer asynchronously, causing the offset to be not calculated when the dialog was just added to the DOM.

In addition the `contentBoxSize` was observed to report the width of the element without any borders. We are able to simplify the code by querying the client rect instead which reports the proper width.

Fixes #5519