Use DOMDocument + textContent to strip HTML tags for search
This is likely going to result in better (i.e. more correct) results compared
to some naive tag stripping. As an example, DOMDocument will correctly
understand and handle HTML entities, specifically `&lt;` and `&gt;` that *need*
to be encoded for use within HTML.

Updating the search index is a somewhat expensive operation anyway, making it
worth the overhead.

see #5558