Fix layout shifts caused by early rendering passes
The previous CSS code relied on `.content:not(:last-child)` to reserve the space required for the right sidebar.

This selector does not work as expected when the browser does an early rendering pass while the `.content` block is still open. It causes the space for the sidebar to be reserved but once the parsing continues the browser realizes that there is no sidebar (=`.content` is the last child) and layout is visually updated.

See https://www.woltlab.com/community/thread/301233-template-list-behavior-when-sorting-results/